+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 0.5.47 (Pending):
+* Upgrade to compile ad test against GWT 2.7.0.
 
 ## 0.5.46:
 * Add CollectorEntityChangeListener that records entity change events, useful during testing.