+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## 0.5.47 (Pending):
-* Upgrade to compile ad test against GWT 2.7.0.
+* Upgrade to compile and test against GWT 2.7.0.
 * Add field_filter dependency to expand the capabilities of the token service.
 
 ## 0.5.46: